menülerde öğretmenler için düzenleme, hareketler eklendi Simple/style.css menü renkleri düzenlendi kurulum sql içine eo_friends eklendi

+++ dataRatingList.php
@@ -143,7 +143,7 @@ function delWithCon(deletepage_url,field_value,messagetext) {
                 <h2 class="PostHeaderIcon-wrapper"> <span class="PostHeader"><img src="img/logo1.png" border="0" style="vertical-align: middle;" alt="main" title="<?php echo $metin[286]?>"/> - <?php echo $metin[287]?> </span> </h2>
                 <div class="PostContent">
                   <?php
-	if ($tur=="2")	{
+	if ($tur=="2" || $tur=="1")	{
 	 //
 $currentPage = $_SERVER["PHP_SELF"];
 
@@ -152,7 +152,7 @@ if (isset($_SERVER['QUERY_STRING'])) {
   $editFormAction .= "?" . htmlentities($_SERVER['QUERY_STRING']);
 }
 
-if ((isset($_GET['id'])) && ($_GET['id'] != "") && ($_GET['delCon'] == "1")) {
+if ((isset($_GET['id'])) && ($_GET['id'] != "") && ($_GET['delCon'] == "1") && $tur=="2") {
   $deleteSQL = sprintf("DELETE FROM eo_rating WHERE id=%s",
                        GetSQLValueString($_GET['id'], "int"));
 
@@ -191,7 +191,7 @@ $arayici =  temizle($_GET['arama']);
 						LEFT OUTER JOIN eo_4konu ON eo_rating.konuID = eo_4konu.id 
 						";
 
-if(!empty($_POST["sil"]) && $_POST["silIzin"]=="evet") {
+if(!empty($_POST["sil"]) && $_POST["silIzin"]=="evet" && $tur=="2") {
    $silinenler = "''";
    for ($i = 0; $i < count($_POST["sil"]); $i++)
       $silinenler .= ",'".$_POST["sil"][$i]."'";
@@ -314,10 +314,13 @@ if ($totalRows_eoUsers>0)
 					    echo $row_eoUsers['value'];   
 					  ?></td>
                         <td nowrap="nowrap" <?php echo "style=\"background-color: $row_color;\""?>><?php echo tarihOku2($row_eoUsers['rateDate']); ?></td>
+                        <?php  if($tur=="2") { ?>
                         <td align="center" nowrap="nowrap" valign="middle" ><a href="#" onclick="javascript:delWithCon('<?php echo $currentPage;?>',<?php echo $row_eoUsers['id']; ?>,'<?php echo $metin[104]?>');"><img src="img/cross.png" alt="delete" width="16" height="16" border="0" style="vertical-align: middle;"  title="<?php echo $metin[102]?>"/></a> |
                           <input type="checkbox" name="sil[]" id="kayitSecici<?php echo $row_eoUsers['id']; ?>" value="<?php echo $row_eoUsers['id']; ?>" /></td>
+                        <?php  } ?>
                       </tr>
                       <?php } while ($row_eoUsers = mysql_fetch_assoc($eoUsers)); ?>
+                      <?php  if($tur=="2") { ?>
                       <tr>
                         <td colspan="6" align="center" valign="middle" class="tabloAlt" ><label>
                             <input name="tumunuSec" type="checkbox" id="tumunuSec" onclick="javascript: 
@@ -337,6 +340,7 @@ if ($totalRows_eoUsers>0)
                             <input type="submit" name="Sil" id="Sil" value="<?php echo $metin[37]?>" />
                           </label></td>
                       </tr>
+                      <?php  } ?>
                     </table>
                   </form>
                   <?php
